Throw hands up in despair
Using arrays of pointers to keep stuff sorted is not an option if said stuff is being reallocated. Not even pointer arithmetic can help!

+++ neat.h
@@ -4,7 +4,7 @@
 
 typedef struct Individual {
 	CPPN genotype;
-	int score;
+	double score;
 	unsigned int species_id;
 } Individual;
